Move actions to separate module
The actions are only used by the 'new' code. `swap_protocols` is tied
to rfc003. If we move the actions into their own module (and export
them into the root cnd module) it will make removal of rfc003 slightly
easier.
